+++ parser/src/azureVmPricingConfig.ts
@@ -1,0 +1,6 @@
+export interface AzureVmPricingConfig {
+  culture: string;
+  currency: string;
+  operatingSystem: string;
+  region: string;
+}
