+++ parser/src/logMessage.ts
@@ -1,0 +1,5 @@
+export interface LogMessage {
+  loggedAt: Date;
+  level: 'warn' | 'error';
+  args: any[];
+}
